cloud_functions/search_candidates/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.13.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
openai==1.6.1
google-generativeai==0.3.2
aiohttp==3.9.1
tenacity==8.2.3